requirements.txt
beautifulsoup4==4.9.1
dj-static==0.0.6
Django==2.2.9
django-bootstrap4==2.2.0
django-stdimage==5.1.1
gunicorn==20.0.4
Pillow==7.2.0
progressbar2==3.53.1
psycopg2-binary==2.8.6
python-utils==2.4.0
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
static3==0.7.0
